fix: move config to module & avoid use lodash template
